+++ debian/copyright
@@ -1,0 +1,27 @@
+Format: https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/
+Upstream-Name: urlinsane
+Source: https://github.com/rangertaha/urlinsane
+
+Files: *
+Copyright: 2018-2024 Rangertaha
+License: GPL-3+
+
+Files: debian/*
+Copyright: 2018-2024 Rangertaha <rangertaha@gmail.com>
+License: GPL-3+
+
+License: GPL-3+
+ Licensed under the Apache License, Version 2.0 (the "License"); you may
+ not use this file except in compliance with the License.  You may obtain
+ a copy of the License at
+ .
+       http://www.apache.org/licenses/LICENSE-2.0
+ .
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ implied.  See the License for the specific language governing
+ permissions and limitations under the License.
+ .
+ On Debian systems, the complete text of the Apache version 2.0 License
+ can be found in "/usr/share/common-licenses/Apache-2.0".
