+++ debian/gbp.conf
@@ -1,0 +1,10 @@
+# [DEFAULT]
+# debian-branch = kali/master
+# debian-tag = kali/%(version)s
+# pristine-tar = True
+
+# [pq]
+# patch-numbers = False
+
+# [dch]
+# multimaint-merge = True
