+++ debian/install
@@ -1,0 +1,1 @@
+/usr/bin/urlinsane
