+++ debian/kali-ci.yml
@@ -1,0 +1,2 @@
+include:
+  - https://gitlab.com/kalilinux/tools/kali-ci-pipeline/raw/master/recipes/kali.yml
