+++ debian/lintian-overrides
@@ -1,0 +1,10 @@
+# spelling errors can't be fixed
+spelling-error-in-binary YuR Your [usr/bin/urlinsane]
+spelling-error-in-binary allready already [usr/bin/urlinsane]
+spelling-error-in-binary lon long [usr/bin/urlinsane]
+spelling-error-in-binary nEAd need [usr/bin/urlinsane]
+spelling-error-in-binary octects octets [usr/bin/urlinsane]
+spelling-error-in-binary securtiy security [usr/bin/urlinsane]
+spelling-error-in-binary singed signed [usr/bin/urlinsane]
+spelling-error-in-binary wTH with [usr/bin/urlinsane]
+spelling-error-in-binary whE when [usr/bin/urlinsane]
