+++ debian/watch
@@ -1,0 +1,4 @@
+version=4
+opts="filenamemangle=s%(?:.*?)?v?(\d[\d.]*)\.tar\.gz%urlinsane-$1.tar.gz%" \
+   https://github.com/rangertaha/urlinsane/archive/refs/tags \
+   (?:.*?/)?v?(\d[\d.]*)\.tar\.gz debian uupdate
